Rework the way locations are handled to use the new locations microformat rather than the old RangeProto.

+++ lib/fluxsource/testpatternfluxsource.cc
@@ -20,7 +20,8 @@ public:
 
         while (fluxmap->duration() < (_config.sequence_length_ms() * 1e6))
         {
-            fluxmap->appendInterval(_config.interval_us() * TICKS_PER_US);
+            fluxmap->appendInterval(
+                _config.interval_us() * (double)TICKS_PER_US);
             fluxmap->appendPulse();
         }
